requirements.txt
kedro==0.17.4
spacy_streamlit==1.0.2
datefinder==0.7.1
joblib==1.0.1
tqdm==4.62.2
matplotlib==3.4.3
plotly==5.3.1
nltk==3.6.2
pandas-datareader==0.10.0
spacy==3.1.2
fastai==2.5.2
sweetviz==2.1.3
hydra-core==1.1.1
scipy==1.7.1
statsmodels==0.12.2
pyfiglet==0.8.post1
gdown==3.13.0
numpy==1.19.5
swifter==1.0.9
pycausalimpact==0.0.16
pandas==1.3.2
texthero==1.1.0
geopy==2.2.0
pytest==6.2.5
pytest-repeat==0.9.1
pytest-benchmark==3.4.1
distfit==1.4.0
pandera==0.7.0
seaborn==0.11.2
newspaper3k==0.2.8
yarl==1.6.3
python-box==5.4.1
snoop==0.4.0
decorator==5.0.9
wordfreq==2.5.1
graphviz==0.17
folium==0.12.1
Faker==8.12.1
heartrate==0.2.1
holidays==0.11.2
autoscraper==1.1.12
textblob==0.15.3
termcolor==1.1.0
array_to_latex==0.82
yellowbrick==1.3.post1
docopt==0.6.2
textacy==0.11.0
dtreeviz==1.3.1
icecream==2.1.1
num2words==0.5.10
maya==0.6.1
pytrends==4.7.3
loguru==0.5.3
pycausalimpact==0.0.16
datapane==0.12.0
ipython==7.27.0
newspaper3k==0.2.8
python-dotenv==0.19.0
rich==10.9.0
scikit_learn==0.24.2
strip_interactive==0.1.5
typer==0.4.0
watermark==2.2.0
interrogate==1.5.0
snscrape==0.3.4
sumy==0.8.1
